api/requirements.txt
flask==2.3.3
flask-cors==4.0.0
flask-socketio==5.3.6
eventlet==0.33.3
python-socketio==5.10.0
python-engineio==4.8.2 